backend2/requirements.txt
psycopg2-binary
redis
PyYAML
python-dotenv
